requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
mariadb>=1.1.8
python-dotenv>=1.0.0
requests>=2.31.0

# FastAPI and web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqladmin>=0.16.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
